+++ src/map/building.cpp
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#include "map_building.hpp"
+#include "map/building.hpp"
 
 #include <math.h>
 #include <stdio.h>
