+++ src/map/dungeon.cpp
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#include "map_dungeon.hpp"
+#include "map/dungeon.hpp"
 
 #include <assert.h>
 #include <fmt/core.h>
