+++ src/map/lightmap.hpp
