+++ src/screen/screen_forest.cpp
@@ -30,8 +30,8 @@
 
 #include "base/entity.hpp"
 #include "main.hpp"
-#include "map_building.hpp"
-#include "map_cell.hpp"
+#include "map/building.hpp"
+#include "map/cell.hpp"
 #include "screen_mainmenu.hpp"
 
 #define FOREST_W 400
